refactor(Profile): 프로필 이미지 회원가입 단계로 수정
