+++ backend/src/main/java/skkuchin/service/ServiceApplication.java
@@ -8,6 +8,7 @@ import org.springframework.scheduling.annotation.EnableScheduling;
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import skkuchin.service.api.dto.UserDto;
+import skkuchin.service.domain.User.Profile;
 import skkuchin.service.domain.User.Major;
 import skkuchin.service.domain.User.Role;
 import skkuchin.service.service.*;
@@ -39,8 +40,8 @@ public class ServiceApplication {
 			userService.saveRole(Role.builder().name("ROLE_USER").build());
 			userService.saveRole(Role.builder().name("ROLE_ADMIN").build());
 
-			userService.saveAdmin(new UserDto.SignUpForm("스꾸친관리자", "admin", "12341234", "12341234",  16, Major.경영학과));
-			userService.saveTestUser(new UserDto.SignUpForm("테스트", "test", "12341234", "12341234", 20, Major.건축학과));
+			userService.saveAdmin(new UserDto.SignUpForm("스꾸친관리자", "admin", "12341234", "12341234",  16, Major.경영학과, Profile.DEFAULT1));
+			userService.saveTestUser(new UserDto.SignUpForm("테스트", "test", "12341234", "12341234", 20, Major.건축학과, Profile.DEFAULT2));
 
 			//String path = System.getProperty("user.dir") + "\\src\\main\\java\\skkuchin\\service\\data\\"; // Window 공통 경로
 			String path = System.getProperty("user.dir") + "/src/main/java/skkuchin/service/data/"; // Mac 공통 경로
