+++ backend/src/main/java/skkuchin/service/domain/User/AppUser.java
@@ -44,6 +44,8 @@ public class AppUser {
     @Enumerated(EnumType.STRING)
     private Campus toggle;
 
+    @Column(nullable = false)
+    @Enumerated(EnumType.STRING)
     private Profile image;
 
     @Enumerated(EnumType.STRING)
